// studio-web/src/api/models/fileInfoDtoDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: v1
 */

export interface FileInfoDtoDTO {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  path?: string | null;
  size?: number;
  /** @nullable */
  mimeType?: string | null;
  uploadedAt?: string;
}